service/requirements.txt
Django==4.0.2
psycopg2==2.9.3
gunicorn==20.1.0
djangorestframework==3.14.0
